requirements.txt
anthropic>=0.8.0
jira>=3.5.1
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.18.0
pillow>=10.2.0
fastapi>=0.110.0
uvicorn>=0.27.1
httpx>=0.27.0
pydantic>=2.6.3 